# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API онлайн-сервиса для IT-школы.",
        "title": "skadi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/solution/for-student": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка решений конкретного ученика.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solution"
                ],
                "summary": "Получение списка решений. [Только ученик]",
                "operationId": "solution-for-student-list",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "HTML",
                        "description": "case-insensitive substring to filter data by task title only for students\nor by task title or student fullname for teacher",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            2
                        ],
                        "description": "solution status IDs (accepted: 1, 2, 3, 4)",
                        "name": "status_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listSolutionOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            }
        },
        "/solution/for-student/{id}": {
            "patch": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Частичное обновление решения (только переданные поля: статус, кроме \"проверено\", ответ, файл ответа) по его id.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solution"
                ],
                "summary": "Обновление решения. [Только ученик]",
                "operationId": "solution-for-student-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID решения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "new status ID",
                        "name": "status_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "new answer",
                        "name": "answer",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "IDs of files to delete from the solution",
                        "name": "delete_files",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "new solution files",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Solution"
                        }
                    },
                    "400": {
                        "description": "неверный статус | отправка на проверку возможна только при наличии ответа"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "решение не найдено"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Contact": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.File": {
            "type": "object",
            "required": [
                "id",
                "mime_type",
                "name",
                "size"
            ],
            "properties": {
                "id": {
                    "description": "file ID",
                    "type": "integer"
                },
                "mime_type": {
                    "description": "file MIME-type",
                    "type": "string"
                },
                "name": {
                    "description": "filename",
                    "type": "string"
                },
                "size": {
                    "description": "file size in bytes",
                    "type": "integer"
                }
            }
        },
        "entity.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pages",
                "per_page",
                "total"
            ],
            "properties": {
                "page": {
                    "description": "current page",
                    "type": "integer",
                    "example": 1
                },
                "pages": {
                    "description": "total pages amount",
                    "type": "integer",
                    "example": 7
                },
                "per_page": {
                    "description": "per page objects",
                    "type": "integer",
                    "default": 10,
                    "example": 5
                },
                "total": {
                    "description": "total objects amount",
                    "type": "integer",
                    "example": 32
                }
            }
        },
        "entity.Profile": {
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/entity.Contact"
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string"
                },
                "id": {
                    "description": "profile id (equals to user id)",
                    "type": "integer"
                },
                "parent_contact": {
                    "$ref": "#/definitions/entity.Contact"
                }
            }
        },
        "entity.Solution": {
            "type": "object",
            "required": [
                "id",
                "status",
                "task"
            ],
            "properties": {
                "answer": {
                    "description": "solution text answer",
                    "type": "string"
                },
                "files": {
                    "description": "solution files",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.File"
                    }
                },
                "grade": {
                    "description": "solution grade",
                    "type": "string"
                },
                "id": {
                    "description": "solution id",
                    "type": "integer"
                },
                "status": {
                    "description": "status object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Status"
                        }
                    ]
                },
                "student": {
                    "description": "student object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "task": {
                    "description": "task object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Task"
                        }
                    ]
                },
                "updated_at": {
                    "description": "last-update datetime of solution",
                    "type": "string"
                }
            }
        },
        "entity.Status": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "description": "status id",
                    "type": "integer"
                },
                "name": {
                    "description": "status name",
                    "type": "string"
                }
            }
        },
        "entity.Task": {
            "type": "object",
            "required": [
                "id",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "task description",
                    "type": "string"
                },
                "files": {
                    "description": "task files",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.File"
                    }
                },
                "id": {
                    "description": "task id",
                    "type": "integer"
                },
                "teacher": {
                    "description": "teacher object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "title": {
                    "description": "task title",
                    "type": "string"
                }
            }
        },
...
        "v1.listSolutionOut": {
            "description": "listSolutionOut represents a solution list data.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "solutions list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Solution"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAccess": {
            "description": "Access JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "access",
            "in": "cookie"
        },
        "JWTRefresh": {
            "description": "Refresh JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "refresh",
            "in": "cookie"
        }
    }
}